requirements.txt
flask-restplus==0.13.0
werkzeug==0.16.1
flask-cors==3.0.8
pymongo[srv]==3.10.1
Flask-PyMongo==2.3.0
mongodb-migrations==0.7.0
requests==2.23.0